CSS template taken from bootstrap also
Just to spice it up and give it the github effect I went with the balck background.

In the template initially there was a lot more than I required and I edited to suit the required website I wanted to form. It has options and drop down menus, but I decided show most of the work using the Search box and java script.