Create profile and search through repos
Repos will go ahead of profile as it will be loading within profile.
This function will go through the repos of the users details from index(0) to the amount. From there I will have an id for which I was want them all to be placed.
The profile will be unique for each github user, the github user is entered by the web user.