User profile made
class: col md 3: This section shows the users Github avatar and has a button that will bring you to their live Github profile page.
Class: col md 9: This creates buttons for giving access to various sectons of the the users page. It also extracts the number of Repos one has, the number of followers and following.
Below this will give the extracted bio and the users location. These details only depend on the users in put and can often say null if they neglect to fill them in.